Drop MR JAR code for <11
Follows up on #83.
