+++ src/main/java/org/wildfly/common/archive/JDKSpecific.java
@@ -12,17 +12,17 @@ import java.util.zip.Inflater;
 final class JDKSpecific {
     private JDKSpecific() {}
 
-    static ByteBuffer inflate(final Inflater inflater, final ByteBuffer[] bufs, final long offset, final int compSize, final int uncompSize) throws DataFormatException, IOException {
+    static ByteBuffer inflate(final Inflater inflater, final ByteBuffer[] bufs, long offset, final int compSize, final int uncompSize) throws DataFormatException, IOException {
         int cnt = 0;
-        byte[] b = new byte[min(16384, compSize)];
         byte[] out = new byte[uncompSize];
         int op = 0;
         while (cnt < compSize) {
             int rem = compSize - cnt;
-            final int acnt = min(rem, b.length);
-            Archive.readBytes(bufs, offset, b, 0, acnt);
-            cnt += acnt;
-            inflater.setInput(b, 0, acnt);
+            final ByteBuffer buf = bufs[Archive.bufIdx(offset + cnt)].duplicate();
+            buf.position(Archive.bufOffs(offset + cnt));
+            buf.limit(min(buf.capacity(), buf.position() + rem));
+            cnt += buf.remaining();
+            inflater.setInput(buf);
             do {
                 op += inflater.inflate(out, op, uncompSize - op);
             } while (! inflater.needsInput());
